Move fluids scenarios
