+++ inductiva/fluids/fluid_tank/fluid_tank_template.splishsplash.json.jinja
