+++ inductiva/fluids/fluid_tank/resources/media/fluid_tank.gif
