--- inductiva/fluids/scenarios/__init__.py
@@ -1,13 +1,0 @@
-#pylint: disable=missing-module-docstring
-from .dam_break import DamBreak
-from .fluid_block import FluidBlock
-from .wind_tunnel import WindTunnel
-from .wind_terrain import WindOverTerrain
-from .fluid_tank import (
-    FluidTank,
-    CubicTankOutlet,
-    CylindricalTankOutlet,
-    CircularTankInlet,
-)
-from .heat_sink import HeatSink
-from ._post_processing import SPHSimulationOutput
